Improve gradle
This does a couple things: 1st, this replaces `modRuntimeOnly` with `modLocalRuntime` which is what we really meant to use anyways (`modLocalRuntime` doesn't contribute runtime dependencies to dependants' runtime configurations). 2nd, this uses the fancy dependency syntax for techreborn as well. 3rd, this makes sure Fabric and FAPI are not commuted to dependents configurations.